Update form.rst
In the example that I modified the text is submitting the form to the existing controller. Specifying the $article variable makes one assume the current controller is the Articles controller, however, the output indicates the controller was the Users controller. In the example, I would think either $article should be replaced with $user or the output should reflect the article controller. The next example in the book indicates a scenario where the method isn't in the current controller.